Fixed (Un)signed warning

+++ Source/DIYCharger/battery.cpp
@@ -200,7 +200,7 @@ void Battery::reset()
   ICharge_ = 0;
 
   // Set all points to 0
-  for (int i = 0; i < sizeof(UBat_)/sizeof(float); ++i)
+  for (size_t i = 0; i < sizeof(UBat_)/sizeof(float); ++i)
   {
     UBat_[i] = 0;
   }
